fixes: visualization and tabs
