--- examples/README.md
@@ -1,5 +1,0 @@
-# Examples
-
-This folder contains extension skeletons showing how to plug in a new dataset and ontology.
-
-See `dataset_plugin_skeleton/` for a minimal example you can copy.
