+++ src/ui/tabs.py
@@ -1,668 +1,16 @@
-#!/usr/bin/env python3
-"""
-converSQL - Natural Language to SQL Query Generation Platform
-Powered by AI, ontological data modeling, and analytical intelligence.
-
-Multi-provider AI support for flexible SQL generation.
-"""
-
 import os
 import time
 
 import pandas as pd
 import streamlit as st
 
-# Import AI service with new adapter pattern
-from src.ai_service import generate_sql_with_ai, get_ai_service
-from src.branding import get_favicon_path, get_logo_data_uri
-
-# Import core functionality
-from src.core import (
-    execute_sql_query,
-    get_ai_service_status,
-    get_analyst_questions,
-    get_table_schemas,
-    scan_parquet_files,
-)
+from src.services.ai_service import generate_sql_with_ai
+from src.services.data_service import display_results
 from src.simple_auth import get_auth_service
+from src.utils import get_analyst_questions
 
-# Import authentication
-from src.simple_auth_components import simple_auth_wrapper
-from src.ui import render_app_footer
-from src.visualization import render_visualization
 
-# Configure page with professional styling
-favicon_path = get_favicon_path()
-
-st.set_page_config(
-    page_title="converSQL - Natural Language to SQL",
-    page_icon=str(favicon_path) if favicon_path else "💬",
-    layout="wide",
-    initial_sidebar_state="expanded",
-)
-
-# Professional CSS styling
-st.markdown(
-    """
-<style>
-:root {
-    --color-background: #FAF6F0;
-    --color-background-alt: #FDFDFD;
-    --color-text-primary: #3A3A3A;
-    --color-text-secondary: #7C6F64;
-    --color-accent-primary: #DDBEA9;
-    --color-accent-primary-darker: #B45F4D;
-    --color-border-light: #E4C590;
-    --color-tag-bg: #E4C590;
-    --color-tag-text: #3A3A3A;
-    --color-footer-bg: #FAF6F0;
-    --color-footer-text: #7C6F64;
-    --color-footer-link-hover: #DDBEA9;
-    --color-success-bg: #F4E8DC;
-    --color-success-border: #E4C590;
-    --color-success-text: #3A3A3A;
-    --color-warning-bg: #F9E7CE;
-    --color-warning-border: #E4C590;
-    --color-warning-text: #7C6F64;
-}
-
-body {
-    background: var(--color-background);
-    color: var(--color-text-secondary);
-}
-
-.main .block-container {
-    padding: 1.25rem 1.75rem 1.75rem 1.75rem;
-    background: var(--color-background);
-    max-width: 1360px;
-    margin: 0 auto;
-}
-
-.stTabs [data-baseweb="tab-list"] {
-    gap: 0.5rem;
-    background: none;
-    border-bottom: 1px solid var(--color-border-light);
-}
-
-.stTabs [data-baseweb="tab"] {
-    background: linear-gradient(160deg, var(--color-background-alt) 0%, var(--color-background) 100%);
-    border: 1px solid var(--color-border-light);
-    border-radius: 8px 8px 0 0;
-    color: var(--color-text-secondary);
-    font-weight: 500;
-    padding: 0.5rem 1rem;
-    margin: 0 0.25rem;
-    border-bottom: none;
-}
-
-.stTabs [aria-selected="true"] {
-    background: linear-gradient(150deg, var(--color-accent-primary) 0%, var(--color-accent-primary-darker) 100%);
-    color: var(--color-background-alt);
-    border-color: var(--color-accent-primary-darker);
-    box-shadow: 0 4px 12px rgba(180, 95, 77, 0.25);
-}
-
-.metric-card {
-    background-color: var(--color-background-alt);
-    border: 1px solid var(--color-border-light);
-    border-radius: 0.5rem;
-    padding: 0.75rem;
-}
-
-.stSelectbox > div > div,
-.stTextArea > div > div {
-    border-radius: 0.5rem;
-    border-color: var(--color-border-light) !important;
-    box-shadow: none !important;
-}
-
-.stButton > button {
-    border-radius: 0.5rem;
-    font-weight: 500;
-    background: var(--color-accent-primary);
-    color: var(--color-text-primary);
-    border: 1px solid var(--color-accent-primary-darker);
-    padding: 0.6rem 1rem;
-}
-
-.stButton > button:hover {
-    background: var(--color-accent-primary-darker);
-    color: var(--color-background-alt);
-}
-
-.stSuccess {
-    background-color: var(--color-success-bg);
-    border-color: var(--color-success-border);
-    color: var(--color-success-text);
-}
-
-.stWarning {
-    background-color: var(--color-warning-bg);
-    border-color: var(--color-warning-border);
-    color: var(--color-warning-text);
-}
-
-div[data-testid="stDataFrame"],
-div[data-testid="stDataFrame"] > div,
-div[data-testid="stDataFrame"] [data-testid="stVerticalBlock"],
-div[data-testid="stDataFrame"] [data-testid="stHorizontalBlock"],
-div[data-testid="stDataFrame"] [data-testid="stDataFrameResizable"] {
-    width: 100% !important;
-    max-width: 100% !important;
-    margin-left: 0 !important;
-    margin-right: 0 !important;
-}
-
-div[data-testid="stDataFrame"] table {
-    width: 100% !important;
-}
-
-div[data-testid="stDataFrame"] [data-testid="StyledDataFrame"] {
-    width: 100% !important;
-    max-width: 100% !important;
-}
-
-.footer-links {
-    margin-top: 1rem;
-    display: flex;
-    justify-content: center;
-    gap: 1.5rem;
-    flex-wrap: wrap;
-}
-
-.footer-links a {
-    color: var(--color-accent-primary-darker);
-    text-decoration: none;
-    font-weight: 500;
-    font-size: 0.85rem;
-}
-
-.footer-links a:hover {
-    color: var(--color-footer-link-hover);
-}
-
-.brand-logo {
-    display: flex;
-    justify-content: center;
-    margin: 0 auto 2.5rem auto;
-    max-width: 640px;
-    padding: 0.5rem 1rem 0 1rem;
-}
-
-.brand-logo img {
-    width: 100%;
-    height: auto;
-    filter: drop-shadow(0 18px 36px rgba(180, 95, 77, 0.22));
-}
-
-
-.sidebar-logo {
-    display: flex;
-    justify-content: center;
-    margin: 0 0 1.25rem 0;
-    padding: 0.35rem 0.75rem 0 0.75rem;
-}
-
-.sidebar-logo img {
-    width: 100%;
-    height: auto;
-    filter: drop-shadow(0 12px 24px rgba(180, 95, 77, 0.18));
-}
-
-
-.sidebar-hero {
-    margin: 0 0 1.5rem 0;
-    padding: 1.35rem 1.2rem 1.4rem 1.2rem;
-    background: var(--color-background-alt);
-    border: 1px solid var(--color-border-light);
-    border-radius: 16px;
-    box-shadow: 0 10px 24px rgba(180, 95, 77, 0.12);
-}
-
-.sidebar-hero__eyebrow {
-    display: block;
-    color: var(--color-text-secondary);
-    text-transform: uppercase;
-    letter-spacing: 0.18em;
-    font-size: 0.7rem;
-    margin-bottom: 0.55rem;
-}
-
-.sidebar-hero__title {
-    color: var(--color-text-primary);
-    font-weight: 500;
-    font-size: 1.35rem;
-    margin: 0 0 0.5rem 0;
-}
-
-.sidebar-hero__subhead {
-    color: var(--color-text-secondary);
-    font-size: 0.92rem;
-    line-height: 1.4;
-    margin: 0 0 0.75rem 0;
-}
-
-.sidebar-hero__pill {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 0.6rem 0.95rem;
-    border-radius: 999px;
-    background: rgba(221, 190, 169, 0.18);
-    border: 1px solid rgba(221, 190, 169, 0.45);
-    font-weight: 500;
-}
-
-.sidebar-hero__pill-label {
-    color: var(--color-text-secondary);
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    font-size: 0.65rem;
-}
-
-.sidebar-hero__pill-value {
-    color: var(--color-accent-primary-darker);
-    font-weight: 600;
-}
-
-.section-card {
-    background: var(--color-background-alt);
-    border: 1px solid var(--color-border-light);
-    border-radius: 18px;
-    padding: 1.25rem 1.25rem;
-    box-shadow: 0 12px 28px rgba(180, 95, 77, 0.10);
-    margin-bottom: 1.25rem;
-}
-
-.section-card__header h3 {
-    color: var(--color-text-primary);
-    font-weight: 400;
-    margin-bottom: 0.5rem;
-}
-
-.section-card__header p {
-    color: var(--color-text-secondary);
-    font-size: 0.97rem;
-    margin: 0;
-}
-
-.text-label {
-    display: block;
-    margin: 1.5rem 0 0.4rem 0;
-    font-weight: 500;
-    color: var(--color-text-primary);
-}
-
-.section-spacer {
-    height: 2rem;
-}
-
-.results-card {
-    background: var(--color-background-alt);
-    border: 1px solid var(--color-border-light);
-    border-radius: 18px;
-    padding: 1.25rem 1.5rem;
-    box-shadow: 0 12px 28px rgba(180, 95, 77, 0.10);
-    margin: 1rem 0 1.5rem 0;
-}
-.results-card div[data-testid="stDataFrame"] {
-    padding-top: 0.5rem;
-}
-</style>
-""",
-    unsafe_allow_html=True,
-)
-
-# Load configuration from environment variables
-DEMO_MODE = os.getenv("DEMO_MODE", "false").lower() == "true"
-
-
-def format_file_size(size_bytes: int) -> str:
-    """Convert bytes to human readable format."""
-    if size_bytes == 0:
-        return "0 B"
-    size_names = ["B", "KB", "MB", "GB", "TB"]
-    import math
-
-    i = int(math.floor(math.log(size_bytes, 1024)))
-    p = math.pow(1024, i)
-    s = round(size_bytes / p, 2)
-    return f"{s} {size_names[i]}"
-
-
-def display_results(result_df: pd.DataFrame, title: str, execution_time: float = None):
-    """Display query results with download option and performance metrics."""
-    if not result_df.empty:
-        # Persist latest results for re-renders and visualization state
-        st.session_state["last_result_df"] = result_df
-        st.session_state["last_result_title"] = title
-
-        st.markdown("<div class='results-card'>", unsafe_allow_html=True)
-        # Compact performance header
-        performance_info = f"✅ {title}: {len(result_df):,} rows"
-        if execution_time:
-            performance_info += f" • ⚡ {execution_time:.2f}s"
-        st.success(performance_info)
-
-        # More compact result metrics in fewer columns
-        col1, col2, col3, col4 = st.columns([2, 2, 2, 3])
-        with col1:
-            st.metric("📊 Rows", f"{len(result_df):,}")
-        with col2:
-            st.metric("📋 Cols", len(result_df.columns))
-        with col3:
-            if execution_time:
-                st.metric("⚡ Time", f"{execution_time:.2f}s")
-        with col4:
-            # Download button in the metrics row to save space
-            csv_data = result_df.to_csv(index=False)
-            filename = title.lower().replace(" ", "_") + "_results.csv"
-            st.download_button(
-                label="📥 CSV",
-                data=csv_data,
-                file_name=filename,
-                mime="text/csv",
-                key=f"download_{title}",
-            )
-
-        # Use full width for the dataframe with responsive height
-        height = min(600, max(200, len(result_df) * 35 + 50))  # Dynamic height based on rows
-        st.dataframe(result_df, use_container_width=True, height=height)
-
-        # Render chart beneath the table
-        render_visualization(result_df)
-
-        st.markdown("</div>", unsafe_allow_html=True)
-        # Mark that we rendered results in this run to avoid double-render in persisted blocks
-        st.session_state["_rendered_this_run"] = True
-    else:
-        st.warning("⚠️ No results found")
-
-
-@st.cache_data(ttl=3600)  # Cache for 1 hour
-def load_parquet_files():
-    """Load and cache parquet files."""
-    return scan_parquet_files()
-
-
-@st.cache_data(ttl=3600)  # Cache for 1 hour
-def load_schema_context(_parquet_files):
-    """Load and cache schema context."""
-    return get_table_schemas(_parquet_files)
-
-
-@st.cache_resource(ttl=3600)  # Cache for 1 hour
-def load_ai_service():
-    """Load and cache AI service with adapter pattern."""
-    return get_ai_service()
-
-
-def initialize_app_data():
-    """Initialize application data and AI services efficiently."""
-    # Initialize session state for non-data items only if missing
-    if "generated_sql" not in st.session_state:
-        st.session_state.generated_sql = ""
-    if "ai_error" not in st.session_state:
-        st.session_state.ai_error = ""
-    if "show_edit_sql" not in st.session_state:
-        st.session_state.show_edit_sql = False
-    # Initialize result persistence slots
-    st.session_state.setdefault("ai_query_result_df", None)
-    st.session_state.setdefault("manual_query_result_df", None)
-    st.session_state.setdefault("last_result_df", None)
-    st.session_state.setdefault("last_result_title", None)
-
-    # Reset per-run flags
-    st.session_state["_rendered_this_run"] = False
-
-    # Check if we need to initialize data (avoid reinitializing on every rerun)
-    if "app_initialized" not in st.session_state or not st.session_state.app_initialized:
-        # Show spinner only if we're actually loading data
-        if "parquet_files" not in st.session_state:
-            with st.spinner("🔄 Loading data files..."):
-                st.session_state.parquet_files = load_parquet_files()
-
-        if "schema_context" not in st.session_state:
-            with st.spinner("🔄 Building schema context..."):
-                st.session_state.schema_context = load_schema_context(st.session_state.parquet_files)
-
-        if "ai_service" not in st.session_state:
-            with st.spinner("🔄 Initializing AI services..."):
-                st.session_state.ai_service = load_ai_service()
-                st.session_state.ai_available = st.session_state.ai_service.is_available()
-
-        # Mark as initialized only after all components are loaded
-        st.session_state.app_initialized = True
-
-
-def main():
-    """Main Streamlit application."""
-
-    # Check if data is available (should be loaded by now)
-    if not st.session_state.get("parquet_files", []):
-        st.error("❌ No data files found. Please ensure Parquet files are in the data/processed/ directory.")
-        return
-
-    logo_data_uri = get_logo_data_uri()
-
-    # Professional sidebar with enhanced styling
-    with st.sidebar:
-        if logo_data_uri:
-            st.markdown(
-                f"""
-        <div class='sidebar-logo'>
-            <img src='{logo_data_uri}' alt='converSQL logo' />
-        </div>
-        """,
-                unsafe_allow_html=True,
-            )
-
-        st.markdown(
-            """
-        <div class='sidebar-hero'>
-            <span class='sidebar-hero__pill-label'>Dataset</span>
-            <span class='sidebar-hero__pill-value'>🏠 Single Family Loan Analytics</span>
-        </div>
-        """,
-            unsafe_allow_html=True,
-        )
-
-        # Professional AI status display (cached)
-        if "ai_status_cache" not in st.session_state:
-            st.session_state.ai_status_cache = get_ai_service_status()
-        ai_status = st.session_state.ai_status_cache
-
-        if ai_status["available"]:
-            provider_name = ai_status["active_provider"].title()
-            st.markdown(
-                """
-            <div style='background-color: var(--color-success-bg); border: 1px solid var(--color-success-border);
-                        border-radius: 6px; padding: 0.75rem; margin: 0.5rem 0;'>
-                <div style='color: var(--color-success-text); font-weight: 500;'>
-                    🤖 AI Assistant: {}
-                </div>
-            </div>
-            """.format(
-                    provider_name
-                ),
-                unsafe_allow_html=True,
-            )
-
-            # AI Provider Selector (if multiple available)
-            ai_service = st.session_state.get("ai_service")
-            if ai_service:
-                available_providers = ai_service.get_available_providers()
-
-                if len(available_providers) > 1:
-                    st.markdown("---")
-                    st.markdown("**🔄 Switch AI Provider:**")
-
-                    provider_options = list(available_providers.keys())
-                    current_provider = ai_service.get_active_provider()
-
-                    # Find current index
-                    default_index = (
-                        provider_options.index(current_provider) if current_provider in provider_options else 0
-                    )
-
-                    selected_provider = st.selectbox(
-                        "Select AI Provider",
-                        options=provider_options,
-                        format_func=lambda x: available_providers[x],
-                        index=default_index,
-                        key="sidebar_provider_selector",
-                        help="Choose which AI provider to use for SQL generation",
-                    )
-
-                    # Update provider if changed
-                    if selected_provider != current_provider:
-                        ai_service.set_active_provider(selected_provider)
-                        st.rerun()
-
-            # Show provider details in professional expander
-            with st.expander("🔧 AI Provider Details", expanded=False):
-                status = ai_status["provider_status"]
-
-                # Show all available providers
-                st.markdown("**Available Providers:**")
-                for provider_key, is_available in status.items():
-                    if provider_key != "active":
-                        provider_display = provider_key.title()
-                        icon = "✅" if is_available else "❌"
-                        status_text = "Available" if is_available else "Unavailable"
-                        active_marker = " **(Active)**" if provider_key == ai_status["active_provider"] else ""
-                        st.markdown(f"- **{provider_display}**: {icon} {status_text}{active_marker}")
-        else:
-            st.markdown(
-                """
-            <div style='background-color: var(--color-warning-bg); border: 1px solid var(--color-warning-border);
-                        border-radius: 6px; padding: 0.75rem; margin: 0.5rem 0;'>
-                <div style='color: var(--color-warning-text); font-weight: 500;'>
-                    🤖 AI Assistant: Unavailable
-                </div>
-                <div style='color: var(--color-warning-text); font-size: 0.85rem; margin-top: 0.25rem;'>
-                    Configure Claude API or Bedrock access
-                </div>
-            </div>
-            """,
-                unsafe_allow_html=True,
-            )
-
-        # Professional configuration status with debug info
-        if DEMO_MODE:
-            st.markdown(
-                """
-            <div style='background-color: var(--color-background-alt); border: 1px solid var(--color-border-light);
-                        border-radius: 6px; padding: 0.5rem; margin: 0.5rem 0;'>
-                <div style='color: var(--color-text-primary); font-size: 0.9rem; font-weight: 500;'>
-                    🧪 Demo Mode Active
-                </div>
-            </div>
-            """,
-                unsafe_allow_html=True,
-            )
-
-            # Show detailed debug information in demo mode
-            auth = get_auth_service()
-
-            with st.expander("🔍 Auth Debug Info", expanded=False):
-                st.markdown("**Authentication Status:**")
-                st.markdown(f"- **Auth Enabled**: {auth.is_enabled()}")
-                st.markdown(f"- **Is Authenticated**: {auth.is_authenticated()}")
-                st.markdown(f"- **Demo Mode**: {DEMO_MODE}")
-
-                # Show current query params
-                query_params = dict(st.query_params)
-                if query_params:
-                    st.markdown("**Current URL Parameters:**")
-                    for key, value in query_params.items():
-                        st.markdown(f"- **{key}**: {str(value)[:100]}")
-                else:
-                    st.markdown("**Current URL Parameters**: None")
-
-                # Show user session info if authenticated
-                if "user" in st.session_state:
-                    user = st.session_state.user
-                    st.markdown("**User Session:**")
-                    st.markdown(f"- **Email**: {user.get('email', 'N/A')}")
-                    st.markdown(f"- **Name**: {user.get('name', 'N/A')}")
-                    st.markdown(
-                        f"- **Auth Time**: {time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(user.get('authenticated_at', 0)))}"
-                    )
-
-                # Configuration status
-                st.markdown("**Configuration:**")
-                st.markdown(f"- **Google Client ID**: {'✅ Set' if os.getenv('GOOGLE_CLIENT_ID') else '❌ Missing'}")
-                st.markdown(
-                    f"- **Google Client Secret**: {'✅ Set' if os.getenv('GOOGLE_CLIENT_SECRET') else '❌ Missing'}"
-                )
-                st.markdown(f"- **Enable Auth**: {os.getenv('ENABLE_AUTH', 'true')}")
-
-        try:
-            st.divider()
-        except AttributeError:
-            st.markdown("<hr />", unsafe_allow_html=True)
-
-        # Professional data tables section
-        with st.expander("📋 Available Tables", expanded=False):
-            parquet_files = st.session_state.get("parquet_files", [])
-            if parquet_files:
-                for file_path in parquet_files:
-                    table_name = os.path.splitext(os.path.basename(file_path))[0]
-                    st.markdown(
-                        f"<div style='color: var(--color-text-primary); margin: 0.25rem 0;'>• <span style='font-weight: 500;'>{table_name}</span></div>",
-                        unsafe_allow_html=True,
-                    )
-            else:
-                st.markdown(
-                    "<div style='color: var(--color-text-secondary); font-style: italic;'>No tables loaded</div>",
-                    unsafe_allow_html=True,
-                )
-
-        # Professional quick stats section
-        with st.expander("📈 Portfolio Overview", expanded=True):
-            if st.session_state.parquet_files:
-                try:
-                    import duckdb  # type: ignore[import-not-found]
-
-                    # Use in-memory connection for stats only
-                    with duckdb.connect() as conn:
-                        # Get record count
-                        total = conn.execute("SELECT COUNT(*) FROM 'data/processed/data.parquet'").fetchone()[0]
-
-                        # Get total file size (cached calculation)
-                        if "total_data_size" not in st.session_state:
-                            st.session_state.total_data_size = sum(
-                                os.path.getsize(f) for f in st.session_state.parquet_files if os.path.exists(f)
-                            )
-                        total_size = st.session_state.total_data_size
-
-                        # Clean metrics display - one per row for readability
-                        st.metric("📊 Total Records", f"{total:,}")
-                        st.metric("💾 Data Size", format_file_size(total_size))
-                        st.metric("📁 Data Files", len(st.session_state.parquet_files))
-                        if total > 0 and total_size > 0:
-                            records_per_mb = int(total / (total_size / (1024 * 1024)))
-                            st.metric("⚡ Record Density", f"{records_per_mb:,} per MB")
-
-                except Exception:
-                    # Fallback stats - clean single column layout
-                    if "total_data_size" not in st.session_state:
-                        st.session_state.total_data_size = sum(
-                            os.path.getsize(f) for f in st.session_state.parquet_files if os.path.exists(f)
-                        )
-                    st.metric("📁 Data Files", len(st.session_state.parquet_files))
-                    st.metric(
-                        "💾 Data Size",
-                        format_file_size(st.session_state.total_data_size),
-                    )
-            else:
-                st.markdown(
-                    "<div style='color: var(--color-text-secondary); font-style: italic; text-align: center; padding: 1rem;'>No data loaded</div>",
-                    unsafe_allow_html=True,
-                )
-
-    # Enhanced tab layout with separate tabs for Manual SQL and Schema
+def render_tabs():
     tab_query, tab_manual, tab_ontology, tab_schema = st.tabs(
         [
             "🔍 Query Builder",
@@ -1210,27 +558,3 @@ def main():
                     st.code(schema_context, language="sql")
                 else:
                     st.warning("Schema not available")
-
-    # Footer content with professional design
-    ai_status = get_ai_service_status()
-    ai_provider_text = ""
-    if ai_status["available"]:
-        provider = ai_status["active_provider"]
-        if provider == "claude":
-            ai_provider_text = "Claude API (Anthropic)"
-        elif provider == "bedrock":
-            ai_provider_text = "Amazon Bedrock"
-        else:
-            ai_provider_text = "AI Assistant"
-    else:
-        ai_provider_text = "Manual Analysis Mode"
-
-    render_app_footer(ai_provider_text)
-
-
-if __name__ == "__main__":
-    # Initialize app data before authentication
-    initialize_app_data()
-
-    # Wrap main function with authentication
-    simple_auth_wrapper(main)()
